requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
python-decouple==3.8
whitenoise==6.6.0
django-widget-tweaks==1.5.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
requests==2.31.0
pandas==2.1.3
django-extensions==3.2.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
dj-database-url==2.1.0